requirements.txt
azure-identity
azure-cognitiveservices-speech
azure-search-documents
python-dotenv
langgraph
langchain
langchain-openai
langchain_openai
langchain_experimental
